+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+### Deprecated
+- Deprecated `craft\commerce\elements\Order::getShippingMethod()`. Use `$shippingMethodName` and `$shippingMethodHandle` instead.
+
 ### Fixed
+- Fixed a bug that occurred when using a custom shipping method. ([#2986](https://github.com/craftcms/commerce/issues/2986))
 - Fixed a bug that could occur when querying products by `type`. ([#2966](https://github.com/craftcms/commerce/issues/2966)) 
 
 ## 3.4.17.2 - 2022-09-16