+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Fixed a bug that occurred when using a custom shipping method. ([#2986](https://github.com/craftcms/commerce/issues/2986))
 - Fixed a bug that could occur when querying products by `type`. ([#2966](https://github.com/craftcms/commerce/issues/2966))
+- Fixed a bug where sharing a cart does not redirect to the correct site url. ([#2990](https://github.com/craftcms/commerce/issues/2990))
 
 ## 4.1.3 - 2022-10-07